Conflicts:
	system/module/ui/lang/zh-cn.php

+++ system/module/ui/model.php
@@ -99,10 +99,13 @@ class uiModel extends model
         if(!$this->file->checkSavePath()) return array('result' => false, 'message' => $this->lang->file->errorUnwritable);
 
         /* Delete old files. */
-        $clientLang = $this->app->getClientLang();
-        $oldFiles = $this->dao->select('id')->from(TABLE_FILE)->where('objectType')->eq($section)->andWhere('lang')->eq($clientLang)->fetchAll('id');
-        foreach($oldFiles as $file) $fileModel->delete($file->id);
-        if(dao::isError()) return array('result' => false, 'message' => $this->lang->fail);
+        if($section != 'logo')
+        {
+            $clientLang = $this->app->getClientLang();
+            $oldFiles = $this->dao->select('id')->from(TABLE_FILE)->where('objectType')->eq($section)->andWhere('lang')->eq($clientLang)->fetchAll('id');
+            foreach($oldFiles as $file) $fileModel->delete($file->id);
+            if(dao::isError()) return array('result' => false, 'message' => $this->lang->fail);
+        }
 
         /* Upload new logo. */
         $uploadResult = $fileModel->saveUpload($htmlTagName);
@@ -119,7 +122,15 @@ class uiModel extends model
         $setting->addedBy   = $file->addedBy;
         $setting->addedDate = $file->addedDate;
 
-        $result = $this->loadModel('setting')->setItems('system.common.site', array($section => helper::jsonEncode($setting)));
+        if($section == 'logo')
+        {
+            $result = $this->loadModel('setting')->setItems('system.common.logo', array($this->getEditingTheme() => helper::jsonEncode($setting)));
+            if($this->post->theme == 'all') $result = $this->loadModel('setting')->setItems('system.common.site', array($section => helper::jsonEncode($setting)));
+        }
+        else
+        {
+            $result = $this->loadModel('setting')->setItems('system.common.site', array($section => helper::jsonEncode($setting)));
+        }
         if($result) return array('result' => true);
 
         return array('result' => false, 'message' => $this->lang->fail);
@@ -753,5 +764,33 @@ class uiModel extends model
 
         if(empty($list)) $this->app->loadClass('zfile')->removeDir(dirname($this->exportPath));
         return $zipFile;
+    }
+
+    /**
+     * Get editing template. 
+     * 
+     * @access public
+     * @return string
+     */
+    public function getEditingTemplate()
+    {
+        $template = $this->config->template->name;
+        if($this->session->editTemplate != '') $template = $this->session->editTemplate;
+        if($this->get->editTemplate != '') $template = $this->get->editTemplate;
+        return $template;
+    }
+
+    /**
+     * Get editing theme. 
+     * 
+     * @access public
+     * @return string
+     */
+    public function getEditingTheme()
+    {
+        $theme = $this->config->template->theme;
+        if($this->session->editTheme != '') $theme = $this->session->editTheme;
+        if($this->get->editTheme != '') $theme = $this->get->editTheme;
+        return $theme;
     }
 }
